refactor: calculator history view
